requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-dotenv
python-jose
PyJWT
pytest
pytest-sqlalchemy
pytest-asyncio
bcrypt
httpx==0.23.0
pydantic
mailersend
cryptography
langchain_openai
langchain_groq
langchain_community
langchain_core
python-multipart